Graded HW3
Nicely done. Your code was very clean, and the error catching was cool!
In future, try to comment your code as you go, it makes it easier to go back and understand. Commit 1-2 example images max to your Git repo as Git is largely reserved for code only.

Close this pull request for +1 points! **4/5**