Graded HW4
Great job included docstrings. Next time include what the function does, function input, and function output. **Documentation: +1**
Code is easy to read with relevant variable names. **Style: +1**
This project seems fairly basic, as we're loading tweets with a specific hashtag to a file and finding the sentiment of the tweet. You could do a lot of interesting analysis with the tweets. **Functionality: +2**

Overall:
Nicely done, I can see that a lot of work went into figuring out how to get twitter working. I docked points off for complexity of the project, but your code is very clean and easy to follow. Great! :D

**Grade 4/5**